bugfix: habilitada opção sair do bloco de notas e renomeados arquivos de previsão do tempo
